test(server): :test_tube: tests for result pattern refactor and missing tests

+++ libs/server/core/application/Features/Publishers/Commands/DeletePublisherCommandHandler.cs
@@ -1,14 +1,21 @@
 namespace Kathanika.Core.Application.Features.Publishers.Commands;
 
-internal sealed class DeletePublisherCommandHandler(IPublisherRepository publisherRepository) : IRequestHandler<DeletePublisherCommand, Result>
+internal sealed class DeletePublisherCommandHandler(
+    IPublisherRepository publisherRepository,
+    IPublicationRepository publicationRepository
+) : IRequestHandler<DeletePublisherCommand, Result>
 {
     public async Task<Result> Handle(DeletePublisherCommand request, CancellationToken cancellationToken)
     {
         Publisher? publisher = await publisherRepository.GetByIdAsync(request.Id, cancellationToken);
         if (publisher is null)
-            return Result.Failure(PublisherAggregateErrors.NotFound(request.Id));
+            return PublisherAggregateErrors.NotFound(request.Id);
 
-        //TODO: var hasPublication = (await publicationRepository.ListAllAsync(x => x.Publisher))
+        bool hasPublication = await publicationRepository
+            .ExistsAsync(x => x.Publisher != null && x.Publisher.Id == request.Id, cancellationToken);
+
+        if (hasPublication)
+            return PublisherAggregateErrors.HasPublication;
         await publisherRepository.DeleteAsync(request.Id, cancellationToken);
         return Result.Success();
     }
